# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * California Housing Price Predictor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface HousingDistrict {
  id: number;
  longitude: number;
  latitude: number;
  housingMedianAge: number;
  totalRooms: number;
  /** @nullable */
  totalBedrooms: number | null;
  population: number;
  households: number;
  medianIncome: number;
  medianHouseValue: number;
  oceanProximity: string;
  roomsPerHousehold: number;
  /** @nullable */
  bedroomsPerRoom: number | null;
  populationPerHousehold: number;
}

export interface HousingStats {
  totalDistricts: number;
  avgMedianHouseValue: number;
  medianMedianHouseValue: number;
  avgMedianIncome: number;
  avgHousingAge: number;
  totalPopulation: number;
  minHouseValue: number;
  maxHouseValue: number;
  nearOceanCount: number;
  inlandCount: number;
  islandCount: number;
  nearBayCount: number;
  lessThanOneHourCount: number;
}

export interface PredictionInput {
  medianIncome: number;
  housingMedianAge: number;
  totalRooms: number;
  totalBedrooms: number;
  population: number;
  households: number;
  latitude: number;
  longitude: number;
  oceanProximity: string;
}

export interface FeatureImportance {
  feature: string;
  importance: number;
}

export interface PredictionResult {
  predictedValue: number;
  confidence: number;
  featureImportance: FeatureImportance[];
  modelAccuracy: number;
}

export interface IncomePricePoint {
  medianIncome: number;
  medianHouseValue: number;
  oceanProximity: string;
  latitude: number;
  longitude: number;
}

export interface RegionPrice {
  oceanProximity: string;
  avgPrice: number;
  medianPrice: number;
  count: number;
}

export interface AgeDistributionBucket {
  ageRange: string;
  count: number;
  avgPrice: number;
}

export interface StarbucksLocation {
  id: number;
  name: string;
  latitude: number;
  longitude: number;
  city: string;
  /** @nullable */
  openYear: number | null;
}

export interface StarbucksPriceImpact {
  openYear: number;
  avgPriceNearby: number;
  avgPriceOther: number;
  percentDiff: number;
  starbucksCount: number;
  label: string;
}

export type GetHousingDistrictsParams = {
  limit?: number;
  offset?: number;
  ocean_proximity?: string;
};