# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: California Housing Price Predictor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /housing/districts:
    get:
      operationId: getHousingDistricts
      tags: [housing]
      summary: Get all housing districts
      description: Returns housing district data from 1990 California census
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: ocean_proximity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of housing districts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HousingDistrict"
...
  /housing/predict:
    post:
      operationId: predictHousingPrice
      tags: [prediction]
      summary: Predict median house value
      description: Uses trained linear regression model to predict housing price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
...
components:
  schemas:
...
    HousingDistrict:
      type: object
      properties:
        id:
          type: integer
        longitude:
          type: number
        latitude:
          type: number
        housingMedianAge:
          type: number
        totalRooms:
          type: number
        totalBedrooms:
          type: ["number", "null"]
        population:
          type: number
        households:
          type: number
        medianIncome:
          type: number
        medianHouseValue:
          type: number
        oceanProximity:
          type: string
        roomsPerHousehold:
          type: number
        bedroomsPerRoom:
          type: ["number", "null"]
        populationPerHousehold:
          type: number
      required:
        - id
        - longitude
        - latitude
        - housingMedianAge
        - totalRooms
        - totalBedrooms
        - population
        - households
        - medianIncome
        - medianHouseValue
        - oceanProximity
        - roomsPerHousehold
        - bedroomsPerRoom
        - populationPerHousehold
...
    PredictionInput:
      type: object
      properties:
        medianIncome:
          type: number
        housingMedianAge:
          type: number
        totalRooms:
          type: number
        totalBedrooms:
          type: number
        population:
          type: number
        households:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        oceanProximity:
          type: string
      required:
        - medianIncome
        - housingMedianAge
        - totalRooms
        - totalBedrooms
        - population
        - households
        - latitude
        - longitude
        - oceanProximity
    PredictionResult:
      type: object
      properties:
        predictedValue:
          type: number
        confidence:
          type: number
        featureImportance:
          type: array
          items:
            $ref: "#/components/schemas/FeatureImportance"
        modelAccuracy:
          type: number
      required:
        - predictedValue
        - confidence
        - featureImportance
        - modelAccuracy
    FeatureImportance:
      type: object
      properties:
        feature:
          type: string
        importance:
          type: number
      required:
        - feature
        - importance
...