# lib/api-zod/src/generated/types/featureImportance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * California Housing Price Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureImportance {
  feature: string;
  importance: number;
}