# lib/api-zod/src/generated/types/housingDistrict.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * California Housing Price Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface HousingDistrict {
  id: number;
  longitude: number;
  latitude: number;
  housingMedianAge: number;
  totalRooms: number;
  /** @nullable */
  totalBedrooms: number | null;
  population: number;
  households: number;
  medianIncome: number;
  medianHouseValue: number;
  oceanProximity: string;
  roomsPerHousehold: number;
  /** @nullable */
  bedroomsPerRoom: number | null;
  populationPerHousehold: number;
}