# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * California Housing Price Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  medianIncome: number;
  housingMedianAge: number;
  totalRooms: number;
  totalBedrooms: number;
  population: number;
  households: number;
  latitude: number;
  longitude: number;
  oceanProximity: string;
}