# lib/api-zod/src/generated/types/predictionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * California Housing Price Predictor API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureImportance } from "./featureImportance";

export interface PredictionResult {
  predictedValue: number;
  confidence: number;
  featureImportance: FeatureImportance[];
  modelAccuracy: number;
}